pyproject.toml
[project]
name = "petstory"
version = "0.1.0"
description = "SaaS para criar histórias de pets com IA"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0,<6.0",
    "pillow>=10.0.0",
    "weasyprint>=60.0",
]
